+++ scripts/actors/ordnance/OrdnanceSheet.js
@@ -30,6 +30,8 @@
  *     embedded-item workflow.
  */
 
+import { DND5E_SHEET_TEMPLATES } from "../dnd5e-compat.js";
+
 const CORE_MODULE_ID  = "causodes-shipcombat-core";
 const DND5E_MODULE_ID = "causodes-shipcombat-dnd5e";
 
@@ -99,7 +101,7 @@ export function buildOrdnanceSheet(NPCActorSheet, ArmorClassConfigApp) {
         scrollable: [""],
       },
       warnings: {
-        template: "systems/dnd5e/templates/actors/parts/actor-warnings-dialog.hbs",
+        template: DND5E_SHEET_TEMPLATES.warnings,
       },
       tabs: {
         id:       "tabs",